+++ src/Client/Dropdown.h
@@ -1,5 +1,10 @@
 #pragma once
 
+#include "Client.h"
+
+class Client;
+class SaveMacroPopup;
+
 #include <Geode/Geode.hpp>
 #include "CCContentSizeTo.h"
 #include "ClippingNode.h"