+++ src/FrameStepper.h
@@ -1,4 +1,4 @@
-#include <Geode/Geode.hpp>
+/*#include <Geode/Geode.hpp>
 //#include <Geode/modify/UILayer.hpp>
 //#include <Geode/modify/PlayLayer.hpp>
 #include "Client/Client.h"
@@ -7,4 +7,4 @@ class FrameStepper
 {
 	public:
 		static inline bool keyDown = false;
-};
+};*/