+++ src/Hacks/Noclip.cpp
@@ -4,11 +4,15 @@
 
 using namespace geode::prelude;
 
-class $modify (PlayLayer)
+class $modify (NoclipLayer, PlayLayer)
 {
     GameObject* ac;
-    //int i = 0;
-    //int d = 0;
+
+    GameObject* last;
+
+    std::vector<GameObject*> dies;
+    int t = 0;
+    int d = 0;
 
     void destroyPlayer(PlayerObject * p0, GameObject * p1)
     {
@@ -17,5 +21,28 @@ class $modify (PlayLayer)
 
         if (!Client::GetModuleEnabled("noclip") || (m_fields->ac == p1))
             PlayLayer::destroyPlayer(p0, p1);
+        else
+        {
+            if (m_fields->last != p1)
+            {
+                m_fields->last = p1;
+                
+                m_fields->d++;
+            }
+
+            if (m_fields->ac != p1)
+                m_fields->t++;
+        }
     }
+
+/*
+    virtual TodoReturn postUpdate(float p0)
+    {
+        PlayLayer::postUpdate(p0);
+
+        if (m_gameState.m_unk1f8 != 0)
+        {
+            log::info("accuracy: {}, {}, {}%", m_fields->t, (m_gameState.m_unk1f8 * 1.0f), (1 - ((m_fields->t * 1.0f) / (m_gameState.m_unk1f8 * 1.0f))) * 100);
+        }
+    }*/
 };